refactor: src.core.config -> src.config
