+++ tests/conftest.py
@@ -23,7 +23,7 @@ dotenv.load_dotenv(dotenv_path=env_path)
 
 @pytest.fixture
 def settings():
-    from src.core.config import get_settings
+    from src.config import get_settings
 
     return get_settings()
